Refactor to move the AccentColorPicker to the components folder (#103)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This change continues the refactor to remove the site-utilities and moves the accent color picker into the components folder.

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
Check the generated link to ensure that the component is still present and rendering correctly.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.
