+++ app/components/accent-color-picker/accent-color-picker.props.ts
