+++ app/components/index.ts
@@ -1,1 +1,2 @@
+export * from "./accent-color-picker";
 export * from "./project-file-transfer";
