+++ app/site-utilities/components/index.ts
@@ -1,4 +1,3 @@
-export * from "./accent-color-picker";
 export * from "./dimension";
 export * from "./direction-switch";
 export * from "./editor";
